Remove protocol NibBased and add instantiate view controller in the protocol NibLoadable
